move test to unit test
